backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt<4.0.0",
    "fastapi>=0.129.0",
    "langchain-groq>=1.1.2",
    "passlib[bcrypt]>=1.7.4",
    "pdfplumber>=0.11.9",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
